+++ packages/shortest/src/utils/errors.ts
@@ -1,0 +1,6 @@
+export class ConfigError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = "ConfigError";
+  }
+}
